move hook into new dir
